+++ netty-codecs/src/main/java/com/kalixia/grapi/codecs/shiro/OAuthAuthorizationServer.java
@@ -1,0 +1,7 @@
+package com.kalixia.grapi.codecs.shiro;
+
+import org.apache.shiro.authc.SimpleAccount;
+
+public interface OAuthAuthorizationServer {
+    SimpleAccount getAccountFromAccessToken(String token);
+}