+++ netty-compiler/src/main/java/com/kalixia/grapi/apt/jaxrs/Options.java
@@ -3,6 +3,7 @@ package com.kalixia.grapi.apt.jaxrs;
 enum Options {
     DAGGER("dagger"),
     METRICS("metrics"),
+    SHIRO("shiro"),
     RXJAVA("rxjava");
 
     private final String value;