+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 This fixes some bugs.
 
+* [Enhancement #4](https://github.com/kalixia/Grapi/issues/4) Scan for Shiro annotations and ensure the security constraints are met
 * [Enhancement #2](https://github.com/kalixia/Grapi/issues/2) Replaced Grapi CORSCodec with new Netty ``` CorsHandler ```
 
 ### Version 0.4.1 ###